+++ branches/trunk/Kernel/spot-on-mailer.cc
@@ -352,9 +352,10 @@ void spoton_mailer::slotRetrieveMail
   dateTime.setTimeSpec(Qt::UTC);
   now.setTimeSpec(Qt::UTC);
 
-  int secsTo = qAbs(now.secsTo(dateTime));
+  qint64 secsTo = qAbs(now.secsTo(dateTime));
 
-  if(!(secsTo <= spoton_common::MAIL_TIME_DELTA_MAXIMUM))
+  if(!(secsTo <= static_cast<qint64> (spoton_common::
+				      MAIL_TIME_DELTA_MAXIMUM)))
     {
       spoton_misc::logError
 	(QString("spoton_mailer::slotRetrieveMail(): "
